refactor: content -> productContent, CartContent로 이름 변경
