--- app/src/main/java/woowacourse/shopping/data/dto/product/Content.kt
@@ -1,18 +1,0 @@
-package woowacourse.shopping.data.dto.product
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Content(
-    @SerialName("category")
-    val category: String,
-    @SerialName("id")
-    val id: Long,
-    @SerialName("imageUrl")
-    val imageUrl: String,
-    @SerialName("name")
-    val name: String,
-    @SerialName("price")
-    val price: Int,
-)
