+++ app/src/main/java/woowacourse/shopping/data/dto/product/ProductsResponse.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ProductsResponse(
     @SerialName("content")
-    val content: List<Content>,
+    val content: List<ProductContent>,
     @SerialName("empty")
     val empty: Boolean,
     @SerialName("first")
